Remove unnecessary walkthrough-steps file
