+++ libs/@hashintel/petrinaut/src/ui/components/walkthrough/index.ts
@@ -1,5 +1,7 @@
 export { WalkthroughContext } from "./walkthrough-context";
-export type { WalkthroughContextValue } from "./walkthrough-context";
+export type {
+  WalkthroughContextValue,
+  WalkthroughStep,
+} from "./walkthrough-context";
 export { WalkthroughProvider } from "./walkthrough-provider";
 export { WalkthroughDialog } from "./walkthrough-dialog";
-export type { WalkthroughStep } from "./walkthrough-steps";
