--- libs/@hashintel/petrinaut/src/ui/components/walkthrough/walkthrough-steps.tsx
@@ -1,7 +1,0 @@
-export type WalkthroughStep = {
-  id: string;
-  title: string;
-  body: React.ReactNode;
-  videoHref: string;
-  videoAlt: string;
-};
